+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Add name editing support
 - Add tabs support
 - Add support for document types, tabs and properties localization
+- Implement EmbeddedPublishedContent Parent property
 - FIX: Return single item from value converter if max items is 1
 
 ## 0.1.0 - 2016-06-21